backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
gunicorn==22.0.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
jsonschema==4.23.0
SQLAlchemy==2.0.36
pyodbc==5.1.0
azure-storage-blob==12.24.0
SQLAlchemy==2.0.36
pyodbc==5.1.0
azure-storage-blob==12.24.0
